Multi-part Coxa
This revision of the coxa prints in three parts, allowing for the use of a 5x11x4mm bearing on the servo back, as well as providing more material for the horn screws to bite into on the servo horn side. The strut attaches to the two outer pieces via 3mm screws.

This model is sized for the Feetech servos, which seem to be 3mm deeper than the servos used in the "stock" build.